% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a map of all the accidents in a given state and year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The numeric code of the desired state (can be a charecter string, e.g. '4')}

\item{year}{A string or number of the year to be used}
}
\value{
The function returns a map (techically a side effect of the function) of all accidents in the given state and year.
}
\description{
This function takes a state code and year (both can be either a string or number).  It then reads in the accident data for that year and creates a plot
of the accidents from the given state in that year.  In addition to the error checking of the helper function fars_read, fars_map_state also checks to
see if the state code is valid and if there are accidents in the given state.  If either of those conditions is not met, the function returns an error.
In additon to the importing make_filename and fars_read, and thus the latter's importing of functions from dplyr and readr, fars_map_state imports
points from graphics (part of base R) and map from maps.
}
\examples{
fars_map_state(4, 2013)
fars_map_state(1, 2013)

}

