% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in a csv file as a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A Character string of the file's location.}
}
\value{
The function returns a tibble containing the data in the file.
}
\description{
The function checks if the file exists and returns a warning if it doens't. If the file does exist, it is read in and converted to a tibble. The function
relies on the read_csv and tbl_df functions from the readr and dplyr packages, respectively. The function assumes that the data is formatted as a csv. If
it isn't, the function may return an error
}
\examples{

}

