% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract month values for given year(s)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years to pass to make_filename.}
}
\value{
The function returns a list of tibbles containing the month values for each year's accident dataset along with that year.
}
\description{
This function takes a vector of years and returns a list of tibbles, one tibble for each year.  Each tibble contains the year and the month value for the
accident dataset for that year.  This function relies on the make_filename and fars_read function.  The latter function depends on functions from
the dplyr and readr packages.  The fars_read_years also imports the mutate and slect functions from dplyr.The function checks, for year item in the
vector of years, that a file with the correct name exists.  If not, it returns a warning.
}
\examples{
fars_read_years(2013)
fars_read_years(2013:2015)
fars_read_years(c("2013","2014"))
}

