% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize accident data with number of observations for each month.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years to pass to fars_read_years.}
}
\value{
The function returns a tibble containing the number of observations for each month (rows)/year (column) combination.
}
\description{
This function takes a vector of years and returns a tidy tibble containing the number of observations in the accident data for each month/year combination.  In
addition to the dependancies of the fars_read_years function, fars_summarize_years imports two functions from dplyr (group_by and summarize) and one
from tidyr (spread).  If an input for year doesn't have a corresponding file, a warning is return (thanks to helper function fars_read_years).
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(2013:2015)
fars_summarize_years(c("2013","2014"))

}

