% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename for annual accident data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or number of the year to be used in the filename}
}
\value{
The function returns a string of a filename for a file contain accident data of the given year.
}
\description{
This function takes a year (in the form of a string or number) and returns a filename for a csv.bs2 file containing accident data for that year.
Note that the function does not return the file, just a string that can be passed to a write function. While the function retruns errors if the input is
the wrong type, is doesn't check to see if the input is a four-digit year.  make_filename is not dependant on any functions other than ones in base R.
}

